pyproject.toml
[tool.poetry]
name = "steal"
version = "0.1.1"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.10.4"
fastapi = "^0.72.0"
uvicorn = {extras = ["standard"], version = "^0.17.0"}
python-dotenv = "^0.19.2"
six = "^1.16.0"
requests = "^2.27.1"
fastapi-auth0 = "^0.3.0"
Jinja2 = "^3.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
